+++ src/mantle/testing/class-block-factory.php
@@ -17,6 +17,7 @@ use function Mantle\Support\Helpers\collect;
  *
  * Used to generate blocks and create presets of blocks for testing.
  *
+ * @method string button(string $text, string $url, array $attributes = [])
  * @method string block(string $name = 'paragraph', ?string $content = null, array $attributes = [])
  * @method string heading(?string $text = null, int $level = 2)
  * @method string image(?string $url = null, ?string $alt = null, array $attributes = [])