+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added a `list()`/`ordered_list()` method to the block factory.
+- Added a `list()`, `ordered_list()`, `reusable()`, and `button()` method to the
+  block factory to generate the corresponding blocks.
 
 ## v1.5.8